fix(labor): days_worked as SUM of shift multipliers, not entry count
Previously days_worked was `COUNT(*)` over priced rows: a half-day
shift counted as 1 day, same as a full-day. A mix of 2 full + 2 half
shifts rendered as 4 days even though the worker actually delivered
3.0 priced days and was paid for 3.0 days. Reported on the Summary
tab as "Huấn shows 88 €/day average when the configured rate is
110 €/day" — the rate denominator (days_worked = 5) was wrong; the
truthful value is 4.0.

Switch the aggregation to SUM(shift_multiplier):
  - full → 1.0
  - half → 0.5
  - overtime → 1.5
  - supplement-only (shift_type IS NULL) → 0
This restores the invariant cost / days_worked == daily_rate for
every non-override row, and gives users a number that reconciles
with reality (e.g. 2.5 days for a 2-full + 1-half month).

Types follow the data:
  - ports.LaborSummaryRow.days_worked: int → Decimal
  - ports.MonthlyWorkerSubRow.days_worked: int → Decimal
  - ports.MonthlyLaborSummaryRow.total_days: int → Decimal
  - DTOs and Pydantic schemas: int → float
The PDF builder gains a small `_format_days` helper that drops
trailing zeros (5 instead of 5.0, 2,5 for 2.5) and the XLSX builder
needs no code change — float values land natively in Excel cells.

Test updates:
  - test_get_monthly_summary_excludes_supplement_only_from_totals:
    March bucket (2 full + 1 half) now asserts 2.5, not 3. The
    existing inline comment already anticipated this.
  - test_get_monthly_summary_nests_per_worker_breakdown:
    March's lone half-day worker now asserts 0.5, not 1.

+++ app/api/v1/labor/schemas.py
@@ -228,7 +228,8 @@ class WorkerSummaryRow(BaseModel):
 
     worker_id: str
     worker_name: str
-    days_worked: int
+    # Fractional priced days — full=1.0, half=0.5, overtime=1.5.
+    days_worked: float
     total_cost: float
     banked_hours: int
     bonus_full_days: int
@@ -240,7 +241,7 @@ class LaborSummaryResponse(BaseModel):
     """Labor summary response."""
 
     rows: List[WorkerSummaryRow]
-    total_days: int
+    total_days: float
     total_cost: float
     total_banked_hours: int
     total_bonus_days: float
@@ -252,7 +253,7 @@ class MonthlyWorkerSubRowResponse(BaseModel):
 
     worker_id: str
     worker_name: str
-    days_worked: int
+    days_worked: float
     total_cost: float
 
 
@@ -265,7 +266,7 @@ class MonthlySummaryRowResponse(BaseModel):
 
     year: int
     month: int
-    total_days: int
+    total_days: float
     total_cost: float
     workers: List[MonthlyWorkerSubRowResponse]
